moving query function to graph engine and updating parser to handle id literals
