+++ helix-engine/src/types.rs
@@ -1,6 +1,8 @@
 use core::fmt;
 use std::{str::Utf8Error, string::FromUtf8Error};
 
+use helixc::parser::parser_methods::ParserError;
+
 #[derive(Debug)]
 pub enum GraphError {
     Io(std::io::Error),
@@ -82,5 +84,11 @@ impl From<Box<bincode::ErrorKind>> for GraphError {
 impl From<libloading::Error> for GraphError {
     fn from(error: libloading::Error) -> Self {
         GraphError::New(error.to_string())
+    }
+}
+
+impl From<ParserError> for GraphError {
+    fn from(error: ParserError) -> Self {
+        GraphError::ConversionError(error.to_string())
     }
 }
